requirements.txt
pandas>=2.0
requests>=2.31
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv>=1.0
Pillow>=10.0
